requirements.txt
python-dateutil==2.8.1
feedparser==5.2.1
python-telegram-bot==12.7
